refactor: Move the arc progress indicator to the core widget as it's going to be the default arc indicator
