+++ lib/src/quiz/conclusion/quiz_conclusion_view.dart
@@ -3,7 +3,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:go_router/go_router.dart';
 import 'package:kana_to_kanji/src/core/widgets/app_scaffold.dart';
 import 'package:kana_to_kanji/src/quiz/conclusion/quiz_conclusion_view_model.dart';
-import 'package:kana_to_kanji/src/quiz/conclusion/widgets/arc_progress_indicator.dart';
+import 'package:kana_to_kanji/src/core/widgets/arc_progress_indicator.dart';
 import 'package:kana_to_kanji/src/quiz/conclusion/widgets/to_review_section.dart';
 import 'package:kana_to_kanji/src/quiz/models/question.dart';
 import 'package:stacked/stacked.dart';
